[IMP] Demo, tests and usability

+++ l10n_br_fiscal/models/document_abstract.py
@@ -2,7 +2,8 @@
 # Copyright (C) 2019  Luis Felipe Mileo - KMEE <mileo@kmee.com.br>
 # License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
 
-from odoo import api, fields, models
+from odoo import api, fields, models, _
+from odoo.exceptions import UserError
 
 from ..constants.fiscal import (
     TAX_FRAMEWORK,
@@ -512,10 +513,18 @@ class DocumentAbstract(models.AbstractModel):
 
     def generate_financial(self):
         for record in self:
+            if self.env.context.get('action_document_confirm'):
+                if (record.amount_missing_payment_value > 0 and
+                        not record.payment_term_id):
+                    raise UserError(
+                        _("O Valor dos lançamentos financeiros é "
+                          "menor que o valor da nota."),
+                    )
+                continue
+
             if record.payment_term_id and self.company_id and self.currency_id:
                 record.financial_ids.unlink()
                 record.fiscal_payment_ids.unlink()
-
                 vals = {
                     'payment_term_id': self.payment_term_id.id,
                     'amount': self.amount_missing_payment_value,
@@ -539,6 +548,7 @@ class DocumentAbstract(models.AbstractModel):
     @api.onchange("fiscal_payment_ids", "payment_term_id")
     def _onchange_fiscal_payment_ids(self):
         financial_ids = []
+
         for payment in self.fiscal_payment_ids:
             for line in payment.line_ids:
                 financial_ids.append(line.id)
